RvSpectML

RvSpectML.NEID.make_manifest

RvSpectML.NEID.read_data

RvSpectML.NEID.read_header

RvSpectML.NEID.read_solar_data

RvSpectML.apply_doppler_factor!

RvSpectML.calc_doppler_factor

RvSpectML.calc_doppler_factor

RvSpectML.calc_normalization

RvSpectML.make_grid_for_chunck

RvSpectML.predict_line_width

RvSpectML.read_manifest

RvSpectML.read_metadata_from_fits

RvSpectML.apply_doppler_factor! — Method

Multiply spectra's λs by dopplerfactor and update spectra metadata, so dopplerfactor knows how to undo the transform.

RvSpectML.calc_doppler_factor — Method

Compute Doppler boost factor (relativistic) for rv and v_perp in km/s

RvSpectML.calc_doppler_factor — Method

Compute Doppler boost factor (non-relativistic) for rv in km/s

RvSpectML.calc_normalization — Method

Calc normalization of spectra based on average flux in a ChunkList.

RvSpectML.make_grid_for_chunck — Method

Create a range with equal spacing between points with end points set based on union of all chunks in timeseries. Takes ChunkListTimeseries, chunk index, and optional an oversample_factor (1).

RvSpectML.predict_line_width — Method

Estimate line width based on stellar Teff (K) and optionally v_rot (km/s). Output in km/s.

RvSpectML.read_manifest — Method

Read manifest containing filename, bjd, target, and optionally additional metadata from CSV file.

RvSpectML.read_metadata_from_fits — Function

Read metadata in FITS header and return data for keys in fields_str/fields as a Dict.

RvSpectML.NEID.make_manifest — Method

Create Dataframe containing filenames and key data for all files neid*.fits in directory

RvSpectML.NEID.read_data — Function

Read NEID (non-solar) data from FITS file, and return in a Spectra2DBasic object.

RvSpectML.NEID.read_header — Method

RvSpectML.NEID.read_solar_data — Function

Read NEID Solar data from FITS file, and return in a Spectra2DBasic object.