RvSpectML Internals

As a heads up, these functions and types are more likely to change going forward than functions and types that are exported.

RvSpectML Internals

Functions

General purpose

Radial Velocity Related

Interpolation

Other

Types

General purpose

Radial Velocity Related

Interpolation

Other

Functions

General purpose

RvSpectML.bin_times — Method

bintimes() Bins times from a SpectralTimeSeriesCommonWavelengths object using the groupings from timeidx in the metadata.

RvSpectML.calc_chunk_rvs_from_taylor_expansion — Method

calc_chunk_rvs_from_taylor_expansion( spectra ) Inputs:

spectra: SpectralTimeSeriesCommonWavelengths

Optional Arguments:

mean: mean spectrum

deriv: dmeanflux/dlnλ

equal_weight: For now, spectra are equal weighted. (true)

Output named pair:

rv: Vector of vector of estimated radial velocities for each chunk

σ_rv: Vector of vector of uncertainties in rv estimates for each chunk

RvSpectML.calc_mean_d2fluxdlnlambda2 — Method

Return mean numerical derivative (d²flux/dlnλ²) based on a common set of wavelengths. Inputs:

flux (2d)

var (2d)

λ (1d)

chunk_map: Array of ranges specifying how each chunk maps into indexes of output derivative

Output:

d²flux/dlnλ²: (1d)

RvSpectML.calc_mean_dfluxdlnlambda — Method

Return mean numerical derivative (dflux/dlnλ) based on a common set of wavelengths. Inputs:

flux (2d)

var (2d)

λ (1d)

chunk_map: Array of ranges specifying how each chunk maps into indexes of output derivative

Output:

dflux/dlnλ: (1d)

RvSpectML.calc_mean_spectrum — Method

Return mean flux (averaging over observations at different times, variance weighted) based on a common set of wavelengths. Inputs: flux & var (2d: pixel, time)

RvSpectML.calc_running_mean_weighted — Method

Smooth array via weighted running mean of n consecutive points

RvSpectML.calc_rvs_from_taylor_expansion — Method

calc_rvs_from_taylor_expansion( spectra ) Inputs:

spectra: SpectralTimeSeriesCommonWavelengths

Optional Arguments:

mean: mean spectrum

deriv: dmeanflux/dlnλ

idx: range of pixel indcies to use for calculation

equal_weight: For now, spectra are equal weighted. (true)

Output named pair:

rv: Vector of estimated radial velocities

σ_rv: Vector of uncertainties in rv estimates

RvSpectML.calc_rvs_from_taylor_expansion_alt — Method

calc_rvs_from_taylor_expansion_alt( spectra ) Experimental version of calcrvsfromtaylorexpansion.

RvSpectML.get_line_shapes — Method

getlineshapes(linelist, clt, fracdepths, vwidth, RVestimate, verbose) For each spectrum in clt, returns arrays for line bisector and width as a function of fractional line depth given in fracdepths. Note that fracdepths must be sorted from largest to smallest. Also returns arrays for line depth, presently a simple minimum over the line's pixels, and valid lines, a boolean array for which lines' shape measurements were calculated successfully. vwidth is the width of the window (in m/s) around each line's center to use. RVestimate (in m/s) is a rough estimate of the star's RV. Note: RVestimate shifts the line centers in linelist to the RV of star, so use RVestimate=0 if linelist is already aligned to star.

RvSpectML.get_line_shapes — Method

getlineshapes(linelist, chunklist, chunkcenters, fracdepths, vwidth, RVestimate, verbose) For a single spectrum in chunklist, returns arrays for line bisector and width as a function of fractional line depth given in fracdepths. Note that fracdepths must be sorted from largest to smallest. Also returns arrays for line depth, presently a simple minimum over the line's pixels, and valid lines, a boolean array for which lines' shape measurements were calculated successfully. vwidth is the width of the window (in m/s) around each line's center to use. RVestimate (in m/s) is a rough estimate of the star's RV. Note: RVestimate shifts the line centers in linelist to the RV of star, so use RVestimate=0 if line_list is already aligned to star.

RvSpectML.get_obs_loc — Method

getobsloc(obs::Symbol) Returns a Dict with long & lat (degrees) and elevation (km) Warning: Currently only has info for :WIYN and :HARPSN.

RvSpectML.make_template_spectra — Method

maketemplatespectra( chunklisttimeseries, [ options ] ) Combine portions of spectra in a ChunkListTimeseries into a template stored as a SpectralTimeSeriesCommonWavelengths. Optionally remove radial velocities before combining spectra based on the :rv_est field in metadata.

Inputs:

chunklisttimeseries

Optional Parametes:

removervest (true)

oversample_factor (1)

smooth_factor (1)

alg (:TemporalGP)

Outputs

matrix: timeseries interpolated to each wavelength in the form of a SpectralTimeSeriesCommonWavelengths

f_mean: mean flux

var_mean: variance of mean flux (WARNING: Not computed accurately yet)

deriv: dflux/dlnλ averaged over spectra

deriv2: d²flux/dlnλ² averaged over spectra

RvSpectML.repack_flux_vector_to_chunk_matrix — Method

repackfluxvectortochunkmatrix(λ, flux, var, chunkmap, λc; alg ) Warning: This doesn't work yet

Radial Velocity Related

RvSpectML.LineFinder.find_line_candidates_in_chunk — Method

find_lines_candidates_in_chunk( chunk, plan ) Convenience function to find links in one chunk of spectrum.

Inputs:

chunk: ChunkOfSpectrum to be analyzed

deriv2: array of second derivative with respect to log λ

Optional Arguments:

plan: LineFinderPlan

Returns DataFrame containing keys:

pixels: array of ranges of pixels wihtin chunk

pixelmaxd2fdlnλ2: pixel with maximum 2nd derivative

RvSpectML.LineFinder.find_lines_in_chunk — Method

findlinesin_chunk( chunk, plan ) Convenience function to find lines in one chunk of spectrum.

Inputs:

chunk

Optional Arguments:

plan: LineFinderPlan

chunkid: If provide non-zero values, then add column `chunkid` to resulting DataFrame

obsid: If provide non-zero values, then add column `obsid` to resulting DataFrame

keepbadfits: If true, then does not filter out non-converged and other bad line fits from resulting dataframe

verbose: If true, prints extra debugging info.

Return:

linefitlist: DataFrame with results of fit to each line

RvSpectML.LineFinder.find_lines_in_chunklist — Method

findlinesinchunklist ( chunklist, linelist ) Convenience function to find lines in each chunk of a spectrum.

Inputs:

chunklist

Optional Arguments:

plan: LineFinderPlan

keepbadfits: If true, then does not filter out non-converged and other bad line fits from resulting dataframe

verbose: If true, prints extra debugging info.

Return:

line_list: DataFrame results of fit to each line

RvSpectML.LineFinder.find_lines_in_chunklist_timeseries — Method

find_lines_in_chunklist_timeseries( chunklist_timeseries, line_list ) Convenience function to find lines in each chunk of each spectrum in a timeseries.

Inputs:

chunklist_timeseries

Optional Arguments:

plan: LineFinderPlan

keepbadfits: If true, then does not filter out non-converged and other bad line fits from resulting dataframe

verbose: If true, prints extra debugging info.

Return:

linelist: Array of results from [findlinesinchunklist](@ref)

RvSpectML.LineFinder.find_pixels_for_all_lines_in_chunklist — Method

find_pixels_for_all_lines_in_chunklist( chunk_list, lines) Return array giving range of pixels for each line within provided chunklist Inputs:

chunk_list

lines: DataFrame containing keys: :fit_min_λ, :fit_max_λ, :chunk_id

Output:

Vector of pixel ranges

RvSpectML.LineFinder.fit_all_lines_in_chunklist_timeseries — Method

fit_line_in_chunklist_timeseries( chunk_list_timeseries, λmin, λmax ) Return DataFrame with results of fits to each line in a chunklisttimeseries (including each observation time) Inputs:

chunklisttimeseries: Data to fit

λmin

λmax

chunk_index: Restricts fitting to specified chunk

Outputs a DataFrame with keys:

fit_a: Normalization of continuum around line

fit_λc: Central wavelength

fit_depth: Line depth

fit_σ²:

fit_b: Slope of continuum around line

fit_covar: Covariance matrix for fit parameters

χ²perdof: quality of fit to chunk

fit_converged: Bool indicating if LsqFit converged

obsidx: index of observation in chunklist_timeseries

chunkid: index of chunk in chunklist_timeseries

pixels: range of pixels that was fit

RvSpectML.LineFinder.fit_line — Method

fitline( λ, flux, var ) Fits a basic Gaussian absorption line times a line (variable slope) to data. Returns a tuple of (param, χ²perdof, isconverged) Warning: Has some hardcoded parameters that likely need to be generalized for different instruments.

RvSpectML.LineFinder.fit_line_in_chunklist_timeseries — Method

fit_line_in_chunklist_timeseries( chunk_list_timeseries, λc, min, λmax, chunk_index) Return DataFrame with results of fits to each line in a given chunk of chunklisttimeseries (including each observation time) Inputs:

chunklisttimeseries: Data to fit

λc

λmin

λmax

chunk_index: Restricts fitting to specified chunk

Outputs a DataFrame with keys:

fit_a: Normalization of continuum around line

fit_λc: Central wavelength

fit_depth: Line depth

fit_σ²:

fit_b: Slope of continuum around line

fit_covar: Covariance matrix for fit parameters

χ²perdof: quality of fit to chunk

fit_converged: Bool indicating if LsqFit converged

obsidx: index of observation in chunklist_timeseries

chunkid: index of chunk in chunklist_timeseries

pixels: range of pixels that was fit

RvSpectML.LineFinder.line_model — Method

line_model( λ, θ ) Calculate model for one Gaussian absorption line evaluated at wavelengths λ Input is an array that will be passed to unpack. Currently assumed to be chunk normalization, central wavelength, depth, σ² and slope of normalization within chunk.

RvSpectML.LineFinder.pack — Method

Utility function for line_model. Don't assume will remain unchanged in future versions.

RvSpectML.LineFinder.unpack — Method

Utility function for line_model. Don't assume will remain unchanged in future versions.

Interpolation

RvSpectML.SincInterpolation.create_filter_curve_orig — Method

Original author: Joe Ninan Converted to Julia and optimized by Christian Gilbertson & Eric Ford Returns a cubit interpolator for windowed sinc Filter curve. noofpoints: number of intepolation points to use in cubic inteprolator

RvSpectML.TemporalGPInterpolation.predict_gp_mean_var — Method

predictgpmeanvar(gp, xpred ; uselogx, use_logy) Inputs:

gp:

xpred: Locations to predict GP at

Optional inputs:

use_logx: If true, apply log transform to xpred before evaluating GP

use_logy: If true, apply exp transform after evaluating GP

Returns tuple with vector of means and variance of GP posterior at locations in xpred.

Other

Types

General purpose

Radial Velocity Related

RvSpectML.LineFinder.LineFinderPlan — Type

Struct containing parameters for running LineFinder

RvSpectML.LineFinder.LineFinderPlan — Method

Create a LineFinderPlan Optional Inputs:

line_width (default 5,000 m/s)

minderiv2 (defaultmin_deriv2)

smoothfactor (defaultsmooth_factor)

minpixelsinline (defaultminpixelsin_line)

uselogλ (defaultuse_logλ),

uselogflux (defaultuse_logflux)

Interpolation

Other